fix dependency security vulnerability by pointing to evergreen CDN version
Responding to email:
GitHub security alert digest
Aqueum’s repository security updates from the week of Mar 26 - Apr 2

	
Aqueum’s personal account
Warning!	
Aqueum / UFS-CatClicker
Known security vulnerabilities detected

Dependency
jquery	
Version
>= 1.2.0 < 3.5.0	
Upgrade to
~> 3.5.0
Defined in
jquery.min.js
Vulnerabilities
CVE-2020-11022 Moderate severity
CVE-2015-9251 Moderate severity
CVE-2020-23064 Moderate severity
CVE-2019-11358 Moderate severity

with help from GitHub Copilot:
Yes, you can use a link that always points to the latest version of jQuery. However, it's important to note that this is generally not recommended for production websites. The reason is that new versions of jQuery can introduce breaking changes, so it's safer to stick to a specific version that you've tested with your site.

That said, if you still want to link to the latest version, you can use the following script tag:
<script src="https://code.jquery.com/jquery-latest.min.js"></script>

This will always load the latest minified version of jQuery. To update this in your HTML file on GitHub:

Navigate to the HTML file in your GitHub repository that contains the script tag.
Click on the pencil icon to edit the file.
Replace the existing jQuery script tag with the one above.
Scroll down to the bottom of the page, write a commit message and click on "Commit changes".